work in progress websocket bridge
